test(admin): move stats fetcher coverage
